Extract `ClassHasAttribute` tests into separate test case
