+++ tests/Framework/ConstraintTest.php
@@ -1402,56 +1402,6 @@ EOF
         $this->fail();
     }
 
-    public function testConstraintClassHasAttribute()
-    {
-        $constraint = Assert::classHasAttribute('privateAttribute');
-
-        $this->assertTrue($constraint->evaluate(\ClassWithNonPublicAttributes::class, '', true));
-        $this->assertFalse($constraint->evaluate(\stdClass::class, '', true));
-        $this->assertEquals('has attribute "privateAttribute"', $constraint->toString());
-        $this->assertCount(1, $constraint);
-
-        try {
-            $constraint->evaluate(\stdClass::class);
-        } catch (ExpectationFailedException $e) {
-            $this->assertEquals(
-                <<<EOF
-Failed asserting that class "stdClass" has attribute "privateAttribute".
-
-EOF
-                ,
-                TestFailure::exceptionToString($e)
-            );
-
-            return;
-        }
-
-        $this->fail();
-    }
-
-    public function testConstraintClassHasAttribute2()
-    {
-        $constraint = Assert::classHasAttribute('privateAttribute');
-
-        try {
-            $constraint->evaluate(\stdClass::class, 'custom message');
-        } catch (ExpectationFailedException $e) {
-            $this->assertEquals(
-                <<<EOF
-custom message
-Failed asserting that class "stdClass" has attribute "privateAttribute".
-
-EOF
-                ,
-                TestFailure::exceptionToString($e)
-            );
-
-            return;
-        }
-
-        $this->fail();
-    }
-
     public function testConstraintClassNotHasAttribute()
     {
         $constraint = Assert::logicalNot(
